+++ src/main/java/uz/o_rustamov/readium/files/repository/AttachmentContentRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @Repository
 public interface AttachmentContentRepository extends JpaRepository<AttachmentContent, Integer> {
 
-    Optional<AttachmentContent> findByAttachment_Id(Integer attachment_id);
+    Optional<AttachmentContent> findByAttachment_Id(long attachment_id);
 
 }
